@@ -18,18 +18,14 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ">=3.10.1 <4.0.0"
+  flutter: ">=3.38.0 <4.0.0" 
 
 dependencies:
   flutter:
     sdk: flutter
   floater:
     path: ../
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
 
 dev_dependencies:
   flutter_test:
