@@ -9,37 +9,22 @@ class ViewTodoPage extends StatefulWidgetBase<ViewTodoPageState> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        title: Text("View Todo"),
-        backgroundColor: this.state.appBarColor,
-      ),
+      appBar: AppBar(centerTitle: true, title: Text("View Todo"), backgroundColor: this.state.appBarColor),
       body: this._buildBody(context),
     );
   }
 
   Widget _buildBody(BuildContext context) {
-    if (this.state.isLoading)
-      return Container(
-        child: Center(
-          child: LoadingSpinner(),
-        ),
-      );
+    if (this.state.isLoading) return Container(child: Center(child: LoadingSpinner()));
 
     return Center(
       child: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: [
-          Text(
-            this.state.todo.title,
-            style: Theme.of(context).textTheme.headline6,
-          ),
+          Text(this.state.todo.title, style: Theme.of(context).textTheme.headlineSmall),
           if (this.state.todo.description != null)
-            Text(
-              this.state.todo.description!,
-              style: Theme.of(context).textTheme.bodyText1,
-            ),
+            Text(this.state.todo.description!, style: Theme.of(context).textTheme.bodyMedium),
         ],
       ),
     );
